requirements.txt
google-generativeai
flask
boto3
aws-wsgi
pip install google-generativeai==0.4.1
